feat: 게시글 삭제 API 개선하기
- DTO로 작성자 정보를 받을 Record 개발함.
- API 엔드포인트 파라미터 변경함.
- deleteBoard 메서드에 verifyBoardOwner 메서드 적용함.
- 회원을 찾는 getMemberEntity 메서드의 로직에서 작성자 정보가 null일 경우를 검증할 수 있도록 수정함.
- 불필요한 예외 상태 제거함.
- 그 외 주석 내용 수정 및 제거함.

+++ src/main/java/com/kakao/saramaracommunity/board/service/BoardServiceImpl.java
@@ -1,6 +1,7 @@
 package com.kakao.saramaracommunity.board.service;
 
 import com.kakao.saramaracommunity.board.dto.business.reqeust.BoardCreateServiceRequest;
+import com.kakao.saramaracommunity.board.dto.business.reqeust.BoardDeleteServiceRequest;
 import com.kakao.saramaracommunity.board.dto.business.reqeust.BoardUpdateServiceRequest;
 import com.kakao.saramaracommunity.board.dto.business.response.BoardCreateResponse;
 import com.kakao.saramaracommunity.board.dto.business.response.BoardGetResponse;
@@ -8,9 +9,9 @@ import com.kakao.saramaracommunity.board.dto.business.response.BoardSearchRespon
 import com.kakao.saramaracommunity.board.entity.Board;
 import com.kakao.saramaracommunity.board.entity.SortType;
 import com.kakao.saramaracommunity.board.exception.BoardBusinessException;
-import com.kakao.saramaracommunity.board.exception.BoardErrorCode;
 import com.kakao.saramaracommunity.board.repository.BoardRepository;
 import com.kakao.saramaracommunity.member.entity.Member;
+import com.kakao.saramaracommunity.member.exception.MemberBusinessException;
 import com.kakao.saramaracommunity.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -19,7 +20,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
+
+import static com.kakao.saramaracommunity.board.exception.BoardErrorCode.BOARD_NOT_FOUND;
+import static com.kakao.saramaracommunity.member.exception.MemberErrorCode.MEMBER_NOT_FOUND;
+import static com.kakao.saramaracommunity.member.exception.MemberErrorCode.UNAUTHORIZED_TO_MEMBER;
 
 @Log4j2
 @Service
@@ -29,23 +35,6 @@ public class BoardServiceImpl implements BoardService {
 
     private final BoardRepository boardRepository;
     private final MemberRepository memberRepository;
-
-    @Override
-    public BoardCreateResponse createBoard(BoardCreateServiceRequest request) {
-        Board createdBoard = boardRepository.save(
-                request.toEntity(
-                        getMemberEntity(request.memberId())
-                )
-        );
-        return BoardCreateResponse.of(createdBoard);
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public BoardGetResponse getBoard(Long boardId) {
-        Board board = getBoardEntity(boardId);
-        return BoardGetResponse.of(board);
-    }
 
     @Override
     @Transactional(readOnly = true)
@@ -63,6 +52,24 @@ public class BoardServiceImpl implements BoardService {
     }
 
     @Override
+    @Transactional(readOnly = true)
+    public BoardGetResponse getBoard(Long boardId) {
+        Board board = getBoardEntity(boardId);
+        return BoardGetResponse.of(board);
+    }
+
+    @Override
+    public BoardCreateResponse createBoard(BoardCreateServiceRequest request) {
+        Board createdBoard = boardRepository.save(
+                request.toEntity(
+                        getMemberEntity(request.memberId())
+                )
+        );
+        log.info("[BoardServiceImpl] 게시글을 등록합니다. 게시글 번호: {}", createdBoard.getId());
+        return BoardCreateResponse.of(createdBoard);
+    }
+
+    @Override
     public void updateBoard(Long boardId, BoardUpdateServiceRequest request) {
         Board savedBoard = getBoardEntity(boardId);
         verifyBoardOwner(savedBoard, request.memberId());
@@ -77,32 +84,27 @@ public class BoardServiceImpl implements BoardService {
         );
     }
 
-    /**
-     * 추후 삭제 로직 내에서 verifyBoardOwner 메소드를 활용해 게시글 작성자 여부를 함께 검사하도록 수정해야 함.
-     */
     @Override
-    public void deleteBoard(Long boardId) {
+    public void deleteBoard(Long boardId, BoardDeleteServiceRequest request) {
         Board savedBoard = getBoardEntity(boardId);
-        // 추후 작성자 검증 필요
+        verifyBoardOwner(savedBoard, request.memberId());
         log.info("[BoardServiceImpl] 게시글을 삭제합니다. 게시글 번호: {}", savedBoard.getId());
         boardRepository.delete(savedBoard);
     }
 
     private void verifyBoardOwner(Board board, Long memberId) {
-        if (!board.getMember().getId().equals(memberId)) throw new BoardBusinessException(BoardErrorCode.UNAUTHORIZED_TO_UPDATE_BOARD);
+        if (!board.getMember().getId().equals(memberId)) throw new MemberBusinessException(UNAUTHORIZED_TO_MEMBER);
     }
 
-    /**
-     * 추후 UNAUTHORIZED_TO_UPDATE_BOARD 예외가 아닌 MemberErrorCode.MEMBER_NOT_FOUND로 수정해야 함.
-     */
     private Member getMemberEntity(Long memberId) {
-        return memberRepository.findById(memberId)
-                .orElseThrow(() -> new BoardBusinessException(BoardErrorCode.UNAUTHORIZED_TO_UPDATE_BOARD));
+        return Optional.ofNullable(memberId)
+                .flatMap(memberRepository::findById)
+                .orElseThrow(() -> new MemberBusinessException(MEMBER_NOT_FOUND));
     }
 
     private Board getBoardEntity(Long boardId) {
         return boardRepository.findById(boardId)
-                .orElseThrow(() -> new BoardBusinessException(BoardErrorCode.BOARD_NOT_FOUND));
+                .orElseThrow(() -> new BoardBusinessException(BOARD_NOT_FOUND));
     }
 
     /**
